IGNORE now works from the profile.
There's just one small quirk on the UI when you click the ignore button
of a package that is ignored already from its profile. Not a
deal-breaker at this point. Lots of unit tests to right to nail this
down but as a proof of concept this is solid. Getting "FOCUS" to work
would require more invasive work.